[Update] 로그인, 세션기능 업데이트 / 게시글 세션으로 업데이트

+++ src/main/java/fashion/look_book/controller/HomeController.java
@@ -16,7 +16,7 @@ public class HomeController {
     //@GetMapping("/home")
     public CreateMemberDto test() {
 
-        Member member = new Member("hi", "1234", "abc", 24, true);
+        Member member = new Member("hi", "1234", "abc", 24L, true);
 
         Long id = memberService.join(member);
 
@@ -32,7 +32,7 @@ public class HomeController {
         private String id;
         private String password;
         private String nickname;
-        private int age;
+        private Long age;
         private boolean sex;
 
         public CreateMemberDto(Member member) {
